Remove redudant files
